AllFine

+++ m-DengYufan/jumpcow/jumpcow.cpp
@@ -4,16 +4,16 @@
 
 using namespace std;
 
-int p, a[150001], f[150001][2];
+LL p, a[150001], f[150001][2];
 
 int main() {
   freopen("jumpcow.in", "r", stdin);
   freopen("jumpcow.out", "w", stdout);
   cin >> p;
-  for (int i = 1; i <= p; i++) {
+  for (LL i = 1; i <= p; i++) {
     cin >> a[i];
   }
-  for (int i = 1; i <= p; i++) {
+  for (LL i = 1; i <= p; i++) {
     f[i][0] = max(f[i - 1][1] - a[i], f[i - 1][0]);
     f[i][1] = max(f[i - 1][0] + a[i], f[i - 1][1]);
   }
